requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
streamlit>=1.30.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
groq>=0.4.1
python-dotenv>=1.0.0
langchain-text-splitters>=0.0.1
